requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
PyJWT>=2.8.0
redis>=5.0.0
locust>=2.28.0
psutil>=5.9.0
structlog>=24.1.0
ruff>=0.6.9
black>=24.4.0
isort>=5.13.2
mypy>=1.10.0
pre-commit>=3.7.1
radon>=6.0.1
mkdocs>=1.6.0
mkdocstrings[python]>=0.25.2
